requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.12.1
idna==3.7
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.5.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
namex==0.0.8
numpy==1.26.3
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.12.1
outcome==1.3.0.post0
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
platformdirs==4.2.2
protobuf==4.25.5
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.1
python-pptx==1.0.2
pytweening==1.0.7
requests==2.31.0
rich==13.8.1
rubicon-objc==0.4.7
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.30
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
virtualenv==20.26.2
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
